Fix downloading achievements for Xbox One games
Xbox achievements only worked for PC games because the GameId value in
Playnite is structured differently for PC games than for console games:
- For PC games, GameId equals to PFN (package family name) value (e.g.
  DevolverDigital.ObservationXB1_6kzv4j18v0c96)
- For console games, GameId is structured as
  CONSOLE_{titleId}_Application (e.g. CONSOLE_2041763016_Application)

This caused the call to GetTitleInfo to fail as it expected a PFN value.
Since the call was only needed to get titleId for a PFN, it's not even
needed for console games. For those, the code now parses titleId from
Playnite's GameId value instead of calling GetTitleInfo.